Fixed FV0.14.0 SQL crash

+++ init/framework/1.14.0.php
@@ -29,21 +29,23 @@ sql_query('CREATE TABLE `servers_ssh_proxies` (`id`         INT(11)     NOT NULL
 /*
  * Inserting proxy relation on new table
  */
-$servers = sql_query('SELECT `id`, `ssh_proxies_id` FROM `servers` WHERE ssh_proxies_id IS NOT NULL');
+if(sql_column_exists('servers', 'ssh_proxies_id')){
+    $servers = sql_query('SELECT `id`, `ssh_proxies_id` FROM `servers` WHERE ssh_proxies_id IS NOT NULL');
 
-while($server = sql_fetch($servers)){
-    sql_query('INSERT INTO `proxy_servers` (`meta_id`, `servers_id`, `proxies_id`)
-               VALUES                      (:meta_id , :servers_id , :proxies_id )',
+    while($server = sql_fetch($servers)){
+        sql_query('INSERT INTO `proxy_servers` (`meta_id`, `servers_id`, `proxies_id`)
+                   VALUES                      (:meta_id , :servers_id , :proxies_id )',
 
-               array(':meta_id'    => meta_action(),
-                     ':servers_id' => $server['id'],
-                     ':proxies_id' => $server['ssh_proxies_id']));
+                   array(':meta_id'    => meta_action(),
+                         ':servers_id' => $server['id'],
+                         ':proxies_id' => $server['ssh_proxies_id']));
+    }
+
+    /*
+     * Removing ssh_proxies_id from servers table
+     */
+    sql_foreignkey_exists ('servers', 'fk_servers_ssh_proxies_id', 'ALTER TABLE `servers` DROP FOREIGN KEY `fk_servers_ssh_proxies_id`');
+    sql_index_exists      ('servers', 'ssh_proxies_id'           , 'ALTER TABLE `servers` DROP INDEX  `ssh_proxies_id`');
+    sql_column_exists     ('servers', 'ssh_proxies_id'           , 'ALTER TABLE `servers` DROP COLUMN `ssh_proxies_id`');
 }
-
-/*
- * Removing ssh_proxies_id from servers table
- */
-sql_foreignkey_exists ('servers', 'fk_servers_ssh_proxies_id', 'ALTER TABLE `servers` DROP FOREIGN KEY `fk_servers_ssh_proxies_id`');
-sql_index_exists      ('servers', 'ssh_proxies_id'           , 'ALTER TABLE `servers` DROP INDEX  `ssh_proxies_id`');
-sql_column_exists     ('servers', 'ssh_proxies_id'           , 'ALTER TABLE `servers` DROP COLUMN `ssh_proxies_id`');
 ?>
